initial commit2

+++ src/main/java/com/bookbouqet/user/TokenRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface TokenRepository extends CrudRepository<Token, Long> {
+public interface TokenRepository extends CrudRepository<Token, Integer> {
    Optional<Token> findByToken(String token);
 }
